Move the knowledge models into their own app
The knowledge base arrived under tickets because shortcuts live there and were
its first source. But it has grown four source types, its own CRUDL, articles
and article images, and only one of those source types has anything to do with
ticketing - so tickets was carrying a feature rather than containing one.

Knowledge, KnowledgeItem, KnowledgeChunk, Article and ArticleImage move to a new
knowledge app, along with their views, forms, templates and tests. Shortcut
stays in tickets, where it belongs: it is a ticketing feature that the knowledge
base happens to index. The knowledge menu still lists shortcuts and still opens
tickets' own create modal, so the split is in ownership, not in the UI.

The migrations are recreated from scratch rather than moved, since none of this
has shipped and there is no data to preserve. The vector extension and its
version guard are hand-written back into the new initial migration, as
makemigrations cannot know about an extension with no model representation. The
shortcut staleness index stays behind in tickets with the model it indexes.

Note that reusing an existing test database across this change fails - the
knowledge_pending index name is unchanged, so it collides on re-creation. A
fresh test database is needed once.
