+++ locale/en_US/LC_MESSAGES/django.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2026-08-04 00:05+0000\n"
+"POT-Creation-Date: 2026-08-04 16:23+0000\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Language: \n"
 "MIME-Version: 1.0\n"
@@ -2749,6 +2749,77 @@ msgstr ""
 msgid "Calls"
 msgstr ""
 
+msgid "Website"
+msgstr ""
+
+msgid "Documents"
+msgstr ""
+
+msgid "The address to crawl, e.g. https://help.example.com"
+msgstr ""
+
+msgid "Max Pages"
+msgstr ""
+
+msgid "Refresh"
+msgstr ""
+
+msgid "Shortcuts"
+msgstr ""
+
+msgid "Helpdesk"
+msgstr ""
+
+msgid "Indexing"
+msgstr ""
+
+msgid "Daily"
+msgstr ""
+
+msgid "Weekly"
+msgstr ""
+
+msgid "Monthly"
+msgstr ""
+
+msgid "Draft"
+msgstr ""
+
+msgid "Published"
+msgstr ""
+
+msgid "New Source"
+msgstr ""
+
+msgid "Upload"
+msgstr ""
+
+msgid "Update Source"
+msgstr ""
+
+msgid "Delete Source"
+msgstr ""
+
+msgid "New Shortcut"
+msgstr ""
+
+msgid "Your knowledge source has been deleted."
+msgstr ""
+
+msgid "Files can only be added to document sets."
+msgstr ""
+
+#, python-format
+msgid "Limit of %d documents reached."
+msgstr ""
+
+msgid "Unsupported file type"
+msgstr ""
+
+#, python-format
+msgid "Limit for file uploads is %s MB"
+msgstr ""
+
 msgid "The name for our alias"
 msgstr ""
 
@@ -2946,13 +3017,6 @@ msgid "Update Label"
 msgstr ""
 
 msgid "Your label has been deleted."
-msgstr ""
-
-msgid "Unsupported file type"
-msgstr ""
-
-#, python-format
-msgid "Limit for file uploads is %s MB"
 msgstr ""
 
 msgid "Channel Disconnected"
@@ -3426,15 +3490,6 @@ msgstr ""
 msgid "Template Fetch Logs"
 msgstr ""
 
-msgid "Daily"
-msgstr ""
-
-msgid "Weekly"
-msgstr ""
-
-msgid "Monthly"
-msgstr ""
-
 msgid "Yearly"
 msgstr ""
 
@@ -3597,44 +3652,14 @@ msgstr ""
 msgid "In Appeal"
 msgstr ""
 
-msgid "Refresh"
-msgstr ""
-
 msgid "Your templates have been fetched and refreshed."
 msgstr ""
 
 msgid "Unable to refresh all templates. See the log for details."
 msgstr ""
 
-msgid "Website"
-msgstr ""
-
-msgid "Documents"
-msgstr ""
-
-msgid "The address to crawl, e.g. https://help.example.com"
-msgstr ""
-
-msgid "Max Pages"
-msgstr ""
-
 #, python-format
 msgid "Teams can have at most %(limit)d topics."
-msgstr ""
-
-msgid "Shortcuts"
-msgstr ""
-
-msgid "Helpdesk"
-msgstr ""
-
-msgid "Indexing"
-msgstr ""
-
-msgid "Draft"
-msgstr ""
-
-msgid "Published"
 msgstr ""
 
 msgid "Closed"
@@ -3647,31 +3672,6 @@ msgid "Unassigned Tickets"
 msgstr ""
 
 msgid "All Tickets"
-msgstr ""
-
-msgid "New Shortcut"
-msgstr ""
-
-msgid "New Source"
-msgstr ""
-
-msgid "Upload"
-msgstr ""
-
-msgid "Update Source"
-msgstr ""
-
-msgid "Delete Source"
-msgstr ""
-
-msgid "Your knowledge source has been deleted."
-msgstr ""
-
-msgid "Files can only be added to document sets."
-msgstr ""
-
-#, python-format
-msgid "Limit of %d documents reached."
 msgstr ""
 
 msgid "New Team"
@@ -5465,6 +5465,53 @@ msgid "No calls"
 msgstr ""
 
 #, python-format
+msgid "You are about to delete the knowledge source <b>%(object)s</b> and everything indexed from it. There is no way to undo this. Are you sure?"
+msgstr ""
+
+msgid "Help articles written by your workspace. Published articles are searchable by AI agents."
+msgstr ""
+
+msgid "No articles"
+msgstr ""
+
+msgid "Update Shortcut"
+msgstr ""
+
+msgid "Delete Shortcut"
+msgstr ""
+
+msgid "No shortcuts"
+msgstr ""
+
+msgid "Canned responses that agents can use to quickly reply to tickets."
+msgstr ""
+
+#, python-format
+msgid "You are about to delete the document <b>%(object.name)s</b> and everything indexed from it. There is no way to undo this. Are you sure?"
+msgstr ""
+
+msgid "Delete Document"
+msgstr ""
+
+msgid "Uploaded documents that AI agents can search when handling conversations."
+msgstr ""
+
+msgid "You have reached the limit of documents for this source and must delete existing ones before you can upload more."
+msgstr ""
+
+msgid "No documents yet"
+msgstr ""
+
+msgid "Last Indexed"
+msgstr ""
+
+msgid "Pages"
+msgstr ""
+
+msgid "No pages yet"
+msgstr ""
+
+#, python-format
 msgid "Aliases for %(name)s"
 msgstr ""
 
@@ -6433,53 +6480,6 @@ msgid "Template"
 msgstr ""
 
 msgid "Other Languages"
-msgstr ""
-
-msgid "Delete Document"
-msgstr ""
-
-msgid "Uploaded documents that AI agents can search when handling conversations."
-msgstr ""
-
-msgid "You have reached the limit of documents for this source and must delete existing ones before you can upload more."
-msgstr ""
-
-msgid "No documents yet"
-msgstr ""
-
-msgid "Last Indexed"
-msgstr ""
-
-msgid "Pages"
-msgstr ""
-
-msgid "No pages yet"
-msgstr ""
-
-#, python-format
-msgid "You are about to delete the knowledge source <b>%(object)s</b> and everything indexed from it. There is no way to undo this. Are you sure?"
-msgstr ""
-
-msgid "Help articles written by your workspace. Published articles are searchable by AI agents."
-msgstr ""
-
-msgid "No articles"
-msgstr ""
-
-msgid "Update Shortcut"
-msgstr ""
-
-msgid "Delete Shortcut"
-msgstr ""
-
-msgid "No shortcuts"
-msgstr ""
-
-msgid "Canned responses that agents can use to quickly reply to tickets."
-msgstr ""
-
-#, python-format
-msgid "You are about to delete the document <b>%(object.name)s</b> and everything indexed from it. There is no way to undo this. Are you sure?"
 msgstr ""
 
 #, python-format
