+++ temba/knowledge/tests/test_knowledgecrudl.py
@@ -6,9 +6,9 @@ from django.core.files.uploadedfile import SimpleUploadedFile
 from django.test.utils import override_settings
 from django.urls import reverse
 
+from temba.knowledge.models import Knowledge, KnowledgeChunk, KnowledgeItem
 from temba.orgs.models import Org
 from temba.tests import CRUDLTestMixin, TembaTest, cleanup
-from temba.tickets.models import Knowledge, KnowledgeChunk, KnowledgeItem
 
 
 class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
@@ -29,7 +29,7 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
 
     @override_settings(ORG_LIMIT_DEFAULTS={"knowledge": 2})
     def test_menu(self):
-        menu_url = reverse("tickets.knowledge_menu")
+        menu_url = reverse("knowledge.knowledge_menu")
 
         # nobody can access if agents feature not enabled
         response = self.requestView(menu_url, self.admin)
@@ -52,13 +52,13 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
 
         # each source links to its read page (results are shortcuts, helpdesk, divider, then the sources)
         menu = self.requestView(menu_url, self.admin).json()["results"]
-        self.assertEqual(reverse("tickets.knowledge_read", args=[website.uuid]), menu[4]["href"])
+        self.assertEqual(reverse("knowledge.knowledge_read", args=[website.uuid]), menu[4]["href"])
 
     def test_read(self):
         website = Knowledge.create_website(self.org, self.admin, "Nyaruka", "https://nyaruka.com")
         docs = Knowledge.create_documents(self.org, self.admin, "Guides")
 
-        read_url = reverse("tickets.knowledge_read", args=[website.uuid])
+        read_url = reverse("knowledge.knowledge_read", args=[website.uuid])
 
         # nobody can access if agents feature not enabled
         response = self.requestView(read_url, self.admin)
@@ -69,9 +69,9 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertRequestDisallowed(read_url, [None, self.agent, self.admin2])
 
         # the system sources aren't served here - they have their own fixed URL pages
-        response = self.requestView(reverse("tickets.knowledge_read", args=[self.system_shortcuts.uuid]), self.admin)
+        response = self.requestView(reverse("knowledge.knowledge_read", args=[self.system_shortcuts.uuid]), self.admin)
         self.assertEqual(404, response.status_code)
-        response = self.requestView(reverse("tickets.knowledge_read", args=[self.system_helpdesk.uuid]), self.admin)
+        response = self.requestView(reverse("knowledge.knowledge_read", args=[self.system_helpdesk.uuid]), self.admin)
         self.assertEqual(404, response.status_code)
 
         # a website read page lists its pages with no upload option
@@ -85,16 +85,16 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertContentMenu(read_url, self.admin, ["Edit", "Delete"])
 
         # a documents read page lists its documents and can upload
-        docs_read_url = reverse("tickets.knowledge_read", args=[docs.uuid])
+        docs_read_url = reverse("knowledge.knowledge_read", args=[docs.uuid])
         response = self.assertReadFetch(docs_read_url, [self.admin], context_object=docs)
         self.assertTrue(response.context["is_documents"])
         self.assertEqual([], list(response.context["items"]))
-        self.assertEqual(reverse("tickets.knowledge_upload", args=[docs.uuid]), response.context["upload_url"])
+        self.assertEqual(reverse("knowledge.knowledge_upload", args=[docs.uuid]), response.context["upload_url"])
         self.assertFalse(response.context["items_limit_reached"])
         self.assertContentMenu(docs_read_url, self.admin, ["Upload", "Edit", "Delete"])
 
     def test_shortcuts(self):
-        shortcuts_url = reverse("tickets.knowledge_shortcuts")
+        shortcuts_url = reverse("knowledge.knowledge_shortcuts")
 
         # nobody can access if agents feature not enabled
         response = self.requestView(shortcuts_url, self.admin)
@@ -120,7 +120,7 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertEqual(404, response.status_code)
 
     def test_helpdesk(self):
-        helpdesk_url = reverse("tickets.knowledge_helpdesk")
+        helpdesk_url = reverse("knowledge.knowledge_helpdesk")
 
         # nobody can access if agents feature not enabled
         response = self.requestView(helpdesk_url, self.admin)
@@ -149,7 +149,7 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertEqual(404, response.status_code)
 
     def test_create(self):
-        create_url = reverse("tickets.knowledge_create")
+        create_url = reverse("knowledge.knowledge_create")
 
         # nobody can access if agents feature not enabled
         response = self.requestView(create_url, self.admin)
@@ -201,7 +201,7 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertEqual(
             {"url": "https://nyaruka.com", "max_depth": 3, "max_pages": 100, "refresh": "weekly"}, website.config
         )
-        self.assertEqual(reverse("tickets.knowledge_read", args=[website.uuid]), response.url)
+        self.assertEqual(reverse("knowledge.knowledge_read", args=[website.uuid]), response.url)
 
         response = self.assertCreateSubmit(
             create_url,
@@ -213,13 +213,13 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
         docs = Knowledge.objects.get(name="Guides")
         self.assertEqual({}, docs.config)
         self.assertEqual(Knowledge.STATUS_READY, docs.status)
-        self.assertEqual(reverse("tickets.knowledge_read", args=[docs.uuid]), response.url)
+        self.assertEqual(reverse("knowledge.knowledge_read", args=[docs.uuid]), response.url)
 
     def test_update(self):
         website = Knowledge.create_website(self.org, self.admin, "Nyaruka", "https://nyaruka.com")
         docs = Knowledge.create_documents(self.org, self.admin, "Guides")
 
-        update_url = reverse("tickets.knowledge_update", args=[website.uuid])
+        update_url = reverse("knowledge.knowledge_update", args=[website.uuid])
 
         # nobody can access if agents feature not enabled
         response = self.requestView(update_url, self.admin)
@@ -231,10 +231,10 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
 
         # neither system source can be updated
         self.assertRequestDisallowed(
-            reverse("tickets.knowledge_update", args=[self.system_shortcuts.uuid]), [self.admin]
+            reverse("knowledge.knowledge_update", args=[self.system_shortcuts.uuid]), [self.admin]
         )
         self.assertRequestDisallowed(
-            reverse("tickets.knowledge_update", args=[self.system_helpdesk.uuid]), [self.admin]
+            reverse("knowledge.knowledge_update", args=[self.system_helpdesk.uuid]), [self.admin]
         )
 
         # website sources expose their crawl settings, document sets just their name
@@ -244,7 +244,7 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
             form_fields={"name": "Nyaruka", "url": "https://nyaruka.com", "max_pages": 500, "refresh": "weekly"},
         )
         self.assertUpdateFetch(
-            reverse("tickets.knowledge_update", args=[docs.uuid]), [self.admin], form_fields={"name": "Guides"}
+            reverse("knowledge.knowledge_update", args=[docs.uuid]), [self.admin], form_fields={"name": "Guides"}
         )
 
         # names must be unique
@@ -268,7 +268,7 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
         website.refresh_from_db()
         self.assertEqual("https://nyaruka.com/docs", website.url)
         self.assertEqual(Knowledge.STATUS_PENDING, website.status)
-        self.assertEqual(reverse("tickets.knowledge_read", args=[website.uuid]), response.url)
+        self.assertEqual(reverse("knowledge.knowledge_read", args=[website.uuid]), response.url)
 
         website.status = Knowledge.STATUS_READY
         website.save(update_fields=("status",))
@@ -291,7 +291,7 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
         )
         self.create_chunk(website, page.uuid, "welcome")
 
-        delete_url = reverse("tickets.knowledge_delete", args=[website.uuid])
+        delete_url = reverse("knowledge.knowledge_delete", args=[website.uuid])
 
         # nobody can access if agents feature not enabled
         response = self.requestView(delete_url, self.admin)
@@ -303,10 +303,10 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
 
         # neither system source can be deleted
         self.assertRequestDisallowed(
-            reverse("tickets.knowledge_delete", args=[self.system_shortcuts.uuid]), [self.admin]
+            reverse("knowledge.knowledge_delete", args=[self.system_shortcuts.uuid]), [self.admin]
         )
         self.assertRequestDisallowed(
-            reverse("tickets.knowledge_delete", args=[self.system_helpdesk.uuid]), [self.admin]
+            reverse("knowledge.knowledge_delete", args=[self.system_helpdesk.uuid]), [self.admin]
         )
 
         response = self.assertDeleteFetch(delete_url, [self.editor, self.admin])
@@ -323,7 +323,7 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
         website = Knowledge.create_website(self.org, self.admin, "Nyaruka", "https://nyaruka.com")
         docs = Knowledge.create_documents(self.org, self.admin, "Guides")
 
-        upload_url = reverse("tickets.knowledge_upload", args=[docs.uuid])
+        upload_url = reverse("knowledge.knowledge_upload", args=[docs.uuid])
 
         # an org that loses the agents feature can't keep uploading to sources it already has
         response = self.requestView(upload_url, self.admin, post_data={})
@@ -342,7 +342,7 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
         # can't upload to a website source or the helpdesk
         for kb in (website, self.system_helpdesk):
             response = self.client.post(
-                reverse("tickets.knowledge_upload", args=[kb.uuid]),
+                reverse("knowledge.knowledge_upload", args=[kb.uuid]),
                 {"file": self.upload(f"{settings.MEDIA_ROOT}/test_media/simple.pdf", "application/pdf")},
             )
             self.assertEqual({"error": "Files can only be added to document sets."}, response.json())
@@ -354,14 +354,14 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertEqual({"error": "Unsupported file type"}, response.json())
 
         # can't upload a file that's too big
-        with patch("temba.tickets.models.KnowledgeItem.MAX_UPLOAD_SIZE", 10):
+        with patch("temba.knowledge.models.KnowledgeItem.MAX_UPLOAD_SIZE", 10):
             response = self.client.post(
                 upload_url, {"file": self.upload(f"{settings.MEDIA_ROOT}/test_media/simple.pdf", "application/pdf")}
             )
             self.assertEqual({"error": "Limit for file uploads is 9.5367431640625e-06 MB"}, response.json())
 
         # can't exceed the documents limit
-        with patch("temba.tickets.models.KnowledgeItem.MAX_DOCUMENTS", 0):
+        with patch("temba.knowledge.models.KnowledgeItem.MAX_DOCUMENTS", 0):
             response = self.client.post(
                 upload_url, {"file": self.upload(f"{settings.MEDIA_ROOT}/test_media/simple.pdf", "application/pdf")}
             )
@@ -391,7 +391,7 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
         )
         self.create_chunk(website, page.uuid, "welcome")
 
-        delete_url = reverse("tickets.knowledgeitem_delete", args=[item.uuid])
+        delete_url = reverse("knowledge.knowledgeitem_delete", args=[item.uuid])
 
         # nobody can access if agents feature not enabled
         response = self.requestView(delete_url, self.admin)
@@ -406,7 +406,7 @@ class KnowledgeCRUDLTest(TembaTest, CRUDLTestMixin):
 
         item_path = item.path
         response = self.assertDeleteSubmit(delete_url, self.admin, object_deleted=item, success_status=302)
-        self.assertEqual(reverse("tickets.knowledge_read", args=[docs.uuid]), response.url)
+        self.assertEqual(reverse("knowledge.knowledge_read", args=[docs.uuid]), response.url)
 
         self.assertEqual(0, docs.chunks.count())
         self.assertFalse(default_storage.exists(item_path))
