+++ temba/settings_common.py
@@ -260,6 +260,7 @@ INSTALLED_APPS = (
     "temba.notifications",
     "temba.flows",
     "temba.tickets",
+    "temba.knowledge",
     "temba.triggers",
     "temba.utils",
     "temba.campaigns",
@@ -372,7 +373,7 @@ PERMISSIONS = {
         "workspace",
     ),
     "request_logs.httplog": ("webhooks",),
-    "tickets.knowledge": ("menu", "upload"),
+    "knowledge.knowledge": ("menu", "upload"),
     "tickets.ticket": ("assign", "menu", "note", "export", "analytics"),
     "triggers.trigger": ("archived", "type", "menu"),
 }
@@ -470,9 +471,9 @@ GROUP_PERMISSIONS = {
         "request_logs.httplog_read",
         "request_logs.httplog_webhooks",
         "templates.template.*",
-        "tickets.article.*",
-        "tickets.knowledge.*",
-        "tickets.knowledgeitem.*",
+        "knowledge.article.*",
+        "knowledge.knowledge.*",
+        "knowledge.knowledgeitem.*",
         "tickets.shortcut.*",
         "tickets.team.*",
         "tickets.ticket.*",
@@ -559,18 +560,18 @@ GROUP_PERMISSIONS = {
         "request_logs.httplog_webhooks",
         "templates.template_list",
         "templates.template_read",
-        "tickets.article_create",
-        "tickets.article_delete",
-        "tickets.article_list",
-        "tickets.article_read",
-        "tickets.article_update",
-        "tickets.knowledge_create",
-        "tickets.knowledge_delete",
-        "tickets.knowledge_menu",
-        "tickets.knowledge_read",
-        "tickets.knowledge_update",
-        "tickets.knowledge_upload",
-        "tickets.knowledgeitem_delete",
+        "knowledge.article_create",
+        "knowledge.article_delete",
+        "knowledge.article_list",
+        "knowledge.article_read",
+        "knowledge.article_update",
+        "knowledge.knowledge_create",
+        "knowledge.knowledge_delete",
+        "knowledge.knowledge_menu",
+        "knowledge.knowledge_read",
+        "knowledge.knowledge_update",
+        "knowledge.knowledge_upload",
+        "knowledge.knowledgeitem_delete",
         "tickets.shortcut_create",
         "tickets.shortcut_delete",
         "tickets.shortcut_list",
