+++ templates/knowledge/knowledge_read.html
@@ -6,9 +6,9 @@
 {% endblock title-text %}
 {% block content %}
   {% if is_website %}
-    {% include "tickets/knowledge/read_website.html" %}
+    {% include "knowledge/read_website.html" %}
   {% else %}
-    {% include "tickets/knowledge/read_documents.html" %}
+    {% include "knowledge/read_documents.html" %}
   {% endif %}
 {% endblock content %}
 {% block extra-script %}
