+++ templates/knowledge/read_website.html
@@ -21,6 +21,6 @@
 </div>
 <table class="list lined scrolled">
   <tbody>
-    {% include "tickets/knowledge/_item_rows.html" with items=items empty_message=_("No pages yet") %}
+    {% include "knowledge/_item_rows.html" with items=items empty_message=_("No pages yet") %}
   </tbody>
 </table>
